Extract the registry code
